general

+++ utils/general.py
@@ -152,12 +152,12 @@ def erase_logs(exp_batch_name):
 
     experiments = os.listdir(os.path.join(root_path, exp_batch_name))
 
-
     for exp in experiments:
         if os.path.isdir(os.path.join(root_path, exp_batch_name, exp)):
             experiments_logs = os.listdir(os.path.join(root_path, exp_batch_name, exp))
             for log in experiments_logs:
-                if not os.path.isdir(os.path.join(root_path, exp_batch_name, exp, log)):
+                if not os.path.isdir(os.path.join(root_path, exp_batch_name, exp, log))\
+                        and '.csv' not in log:
                     os.remove(os.path.join(root_path, exp_batch_name, exp, log))
 
 def erase_wrong_plotting_summaries(exp_batch_name, validation_data_list, ):
@@ -222,7 +222,6 @@ def erase_validations(exp_batch_name, validation_data_list ):
 
     # Get the correct files sizes for each validation
     # open the csv file with the ground_truth
-
 
     for exp in experiments:
         print ("exp", exp)
